LayoutTests/media/video-seek-past-end-paused-expected.txt

Test that seeking paused video past it's duration time sets currentTime to duration and leaves video paused.

RUN(video.src = 'content/test.mp4')
RUN(video.load())

EVENT(load)
EXPECTED (video.paused == 'true') OK
EXPECTED (video.ended == 'false') OK
RUN(video.play())

EXPECTED (video.paused == 'false') OK
EXPECTED (media.currentTime > '0') OK

EXPECTED (media.currentTime == 'media.duration') OK
EXPECTED (video.ended == 'true') OK

END OF TEST
